requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# API & Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
httpx==0.26.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Accounting Integrations
# quickbooks-online==0.8.0  # Uncomment when needed
# xero-python==3.0.0  # Uncomment when needed

# Reports & Visualization
openpyxl==3.1.2
reportlab==4.0.7
matplotlib==3.8.2

# CLI
click==8.1.7
rich==13.7.0

# AI & NLP (for financial insights)
openai==1.6.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
